Убраны неиспользуемые pathVariable, Поиск task приведен в работоспособность. Проверено в postman. PathVariable окончательно вынесены в интерфейс. Оптимизированы URL для поиска task. Обновлён API.md Обновлено изображение схемы БД (один из usr_id исправлен на user_id, редуцированное поле появляться более не должно)

+++ domain/src/main/java/ru/studentsplatform/backend/domain/repository/TaskRepository.java
@@ -21,7 +21,7 @@ public interface TaskRepository extends JpaRepository<Task, Long> {
 
 	@Query("select att from Task att where att.scheduleUserCell.user.id = :userId " +
 			"and att.scheduleUserCell.discipline.semester = :semester")
-	List<Task> findBySemesterByUserId(@Param("userId") Long userId, @Param("semester") Long semester);
+	List<Task> findBySemesterByUserId(@Param("userId") Long userId, @Param("semester") Integer semester);
 
 	@Query("select att from Task att where att.scheduleUserCell.user.id = :userId " +
 			"and att.scheduleUserCell.discipline.subject.id = :subjectId")
